1. font支持外部读取，调整结构以便更好支持外部存取font 2. 调整tt2c和Resource脚本 3. EGUI_CONFIG_PERFORMANCE_LOAD_IMAGE_IN_RAM更换为EGUI_CONFIG_FUNCTION_EXTERNAL_RESOURCE 4. 加入EGUI_FONT_STD_GET_FONT_HEIGHT接口，获取真实高度

+++ src/core/egui_common.c
@@ -96,8 +96,8 @@ void egui_memcpy(void *dest, const void *src, uint32_t n)
     uint8_t *p_dest = (uint8_t*)dest;
     const uint8_t *p_src = (const uint8_t*)src;
     // Check base offset alignment
-    uint8_t base_offset_dest = (int)p_dest & 0x03;
-    uint8_t base_offset_src = (int)p_src & 0x03;
+    uint8_t base_offset_dest = ((uint8_t)p_dest) & 0x03;
+    uint8_t base_offset_src = ((uint8_t)p_src) & 0x03;
     if((base_offset_dest != base_offset_src)
         || (n < 4))
     {
